devstack-manager/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
docker==6.1.3
pyyaml==6.0.1
websockets==12.0
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.0